Add functionality for extracting location information
Split some of extract_zipcode_with_url to  extract_zipcode_with_lat_lng. Add extract_city_and_state_with_lat_lng. Add FBI_Crime_Data.py containing get_crime_data which accepts lat, lng and returns the crime data for the local police department in the form of a dictionary where the key is the crime and the value is the number of times the crime was committed.